mcp/requirements.txt
# ======================================================================
# REQUIREMENTS.TXT - Sistema de Restaurante MCP
# ======================================================================
# Sistema de chatbot inteligente para restaurante usando MCP
# ======================================================================

# ======================================================================
# MCP (MODEL CONTEXT PROTOCOL)
# ======================================================================
mcp[cli]==1.11.0                # Framework oficial para crear servidores MCP

# ======================================================================
# BASE DE DATOS Y ORM
# ======================================================================
pydantic>=2.8.0,<3.0.0          # Validación de datos y serialización (requerido por SQLModel)
sqlmodel==0.0.14                # ORM moderno basado en SQLAlchemy y Pydantic
sqlalchemy==2.0.23              # ORM y toolkit de base de datos
psycopg2-binary==2.9.9          # Driver PostgreSQL para Python

# ======================================================================
# CONFIGURACIÓN Y VARIABLES DE ENTORNO
# ======================================================================
python-dotenv==1.0.0            # Cargar variables de entorno desde archivos .env

# ======================================================================
# AUTENTICACIÓN Y SEGURIDAD
# ======================================================================
python-jose[cryptography]==3.3.0  # JWT tokens y criptografía
passlib[bcrypt]==1.7.4           # Hashing de contraseñas

# ======================================================================
# LOGGING Y MONITOREO
# ======================================================================
structlog==23.2.0               # Logging estructurado

# ======================================================================
# INTELIGENCIA ARTIFICIAL Y LLM
# ======================================================================
# ======================================================================
# COMUNICACIÓN Y APIS EXTERNAS
# ======================================================================
httpx>=0.27                     # Cliente HTTP asíncrono (compatible con mcp)
requests>=2.31.0                # Cliente HTTP simple

# ======================================================================
# UTILIDADES Y HERRAMIENTAS
# ======================================================================
python-dateutil==2.8.2          # Extensiones para datetime

# ======================================================================
# TESTING
# ======================================================================
pytest>=7.4.3                   # Framework de testing
pytest-asyncio>=0.21.1          # Testing asíncrono 

pytz